MAP ERROR FIXED:
Modified leaflet OpenstreetMap by passing in the location list as prop
Passed in the location list prop into the empty useEffect list for re-rendering.
handled the 'Map already initialize' error by destroying the map instance when the component unmounts.
Removed the useState variables previously setting the lat & lon as the lat & lon are now passed in as list prop and looped through to get the lat & lon for each serach location.
Created an App.css file for styling of the application.
Installed dependecies 'style-loader and css-loader' for implementing the App.css stylesheet of the application.
Modified the webpack.config.js file module rules to use the downloader dependencies.