BOOKING FEATURE UPDATED:
Modify the Rest API booking route to check that there is availability at that accommodation for that number of people on the specified date.
Implemented an appropriate HTTP error code if there is no availability or if the date is in the past and included the specific error within the JSON message sent back.
Modified the client side to test the HTTP code returned from the server and display an appropriate user-friendly message to the user.