BOOKING FEATURE UPDATE 1.1:
Modified the checking condition booking route on the server side to query the database and get availabilty for the date and number of people provided by the user.
Added an additional http error checking condition that checks for invalid or empty user input (error 406).
Modified the availability update query to update the availabilty column based on the accommodation ID & the number of people specified when booking is successful.
Updated the client side to handle the response status gotten from the server and display each error appropriately with a user-friendly message.