IMPLEMENTED BOOKING BUTTON FEATURE ON LEAFLET MARKER POPUP:
Added a booking button to the leaflet marker popup feature
Created an input field to collect user input data for booking an accommodation.
Added an eventListener to the button created using leaflet .on map 'popupopen'feature.
created a variable within a bookLocation function that gets the user input value when the button is clicked.
Passed the booking info variable using the lifting up state to a function in the parent component (App.js).
Called the ajaxBook async function to initialize the accommodation booking by querying the appropriate server routes.
Performed various testing to ensure user input data is entered correctly and reading the appropriate http status codes.