DAOs & CONTROLLER FEATURE:
Implemented a dao & a controller feature for the users and locations routes.
Imported the controllers into the respectively & created a controller object & passed in the database connection as an argument.
Handled the get & post requests to route the /users & /locations using the respective controller's method.
CURRENT ISSUE: After login in successfully the location search page does not appear even on page refresh.